requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pyyaml>=6.0
tqdm>=4.60.0
pytest>=7.0.0
